requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
click==7.1.1
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gast==0.3.3
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.3
opencv-python==4.2.0.34
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
plotly==4.6.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Theano==1.0.4
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
wcwidth==0.1.9
Werkzeug==1.0.1
